docs/requirements.txt
jinja2
commonmark
pyyaml